add laravel permission

+++ app/Http/Controllers/TaskController.php
@@ -26,6 +26,7 @@ class TaskController extends Controller
      */
     public function create()
     {
+        $this->authorize('manage tasks');
         return view('tasks.create');
     }
 
@@ -37,6 +38,8 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        $this->authorize('manage tasks');
+
         Task::create($request->validated());
         return redirect()->route('tasks.index');
     }
@@ -49,6 +52,8 @@ class TaskController extends Controller
      */
     public function edit(Task $task)
     {
+        $this->authorize('manage tasks');
+
         return view('tasks.edit', compact('task'));
     }
 
@@ -61,6 +66,8 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        $this->authorize('manage tasks');
+
         $task->update($request->validated());
         return redirect()->route('tasks.index');
     }
@@ -73,6 +80,8 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        $this->authorize('manage tasks');
+        
         $task->delete();
         return redirect()->route('tasks.index');
     }
